I started creating the parent classes "Hero" and "Item".
Oumar, I was having trouble following your idea for how an encounter
would go, so feel free to edit or just comment as to what sort of change
would make for more interesting gameplay.

Anways, this is a good start, and I'm happy to finish up this part of
the project (it's not really necessary at this stage, but might give us
more time later on to perfect the more complicated aspects.